Add updateCounter function to lib/
